remove deprecated method, move test github code out of prod code
